debugged css rules monday afternoon 9-16-13
I had a space between "h1" and ":before" which caused the rule to be
ignored. Also, the underline rule was ignored because the font-family
rule before it didn't have a semicolon after its statement. I added one
and also put the text-decoration:underline; statement on its own line,
and it showed up.

also added a line at the bottom of the servings paragraph:
.servings{border-bottom:2px solid black;"